feat: create window-activity-tracker file
